feat: put tokenTTL in domain pkg
Rem tokenTTL from pkg app and server. Put it into domain pkg. Fix issues associated with this change
